BackEnd/backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Database - MongoDB
motor==3.3.2
pymongo==4.6.1
beanie==1.24.0

# ML & Audio Processing (Python 3.13 compatible)
torch>=2.6.0
transformers>=4.40.0
librosa>=0.10.2
pydub>=0.25.1
scipy>=1.13.0
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.4.0

# Audio Specific
soundfile==0.12.1
webrtcvad==2.0.10
pyaudio==0.2.13

# Utils
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
argon2-cffi==23.1.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Regional Language Support
indic-nlp-library==0.92
indic-transliteration==1.8.1